+++ src/codepost_powertools/__main__.py
@@ -5,9 +5,9 @@ The `codepost_powertools` package on the command line.
 # =============================================================================
 
 import click
+import cloup
 
 from codepost_powertools import grading, rubric
-from codepost_powertools._utils.cli_utils import NaturalOrderCollection
 from codepost_powertools.info import __version__, package_name
 
 # =============================================================================
@@ -16,12 +16,13 @@ _cli_name = "cptools"
 
 # =============================================================================
 
-cli = NaturalOrderCollection(
-    name=_cli_name,
-    sources=[
+
+@cloup.group(
+    _cli_name,
+    sections=[
         # pylint: disable=protected-access
         # Access the `_cli_group.group` group of each submodule
-        group._cli_group.group
+        group._cli_group.group.as_section()
         for group in (
             grading,
             rubric,
@@ -29,13 +30,15 @@ cli = NaturalOrderCollection(
     ],
     help=__doc__,
 )
-
-cli = click.version_option(
+@click.version_option(
     version=__version__,
     package_name=package_name,
     prog_name=_cli_name,
     message="%(prog)s (%(package)s), version %(version)s",
-)(cli)
+)
+def cli():
+    pass
+
 
 # =============================================================================