+++ src/codepost_powertools/grading/_cli_group.py
@@ -4,8 +4,8 @@ Holds the CLI grading group.
 
 # =============================================================================
 
-from codepost_powertools._utils.cli_utils import NaturalOrderGroup
+from codepost_powertools._utils.cli_utils import SectionGroup
 
 # =============================================================================
 
-group = NaturalOrderGroup(name="Grading")
+group = SectionGroup(name="Grading")